feat: add main service

+++ main-service/gateway/src/main/java/ru/practicum/category/controller/AdminCategoryClientController.java
@@ -3,16 +3,14 @@ package ru.practicum.category.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.category.CategoryClient;
 import ru.practicum.dto.CategoryDto;
+import ru.practicum.model.ConstraintViolationException;
 
 import javax.validation.Valid;
-import javax.validation.constraints.Positive;
 
 @Controller
-@Validated
 @RequiredArgsConstructor
 @RequestMapping(path = "/admin/categories")
 public class AdminCategoryClientController {
@@ -25,12 +23,18 @@ public class AdminCategoryClientController {
 
     @PatchMapping("/{catId}")
     public ResponseEntity<Object> patch(@RequestBody @Valid CategoryDto categoryDto,
-                                        @PathVariable @Positive Long catId) {
+                                        @PathVariable Long catId) {
+        if (catId < 1) {
+            throw new ConstraintViolationException("Id should be positive");
+        }
         return client.patch(categoryDto, catId);
     }
 
     @DeleteMapping("/{catId}")
-    public ResponseEntity<Object> delete(@PathVariable @Positive Long catId) {
+    public ResponseEntity<Object> delete(@PathVariable Long catId) {
+        if (catId < 1) {
+            throw new ConstraintViolationException("Id should be positive");
+        }
         return client.delete(catId);
     }
 }
