+++ main-service/gateway/src/main/java/ru/practicum/compilation/controller/AdminCompilationClientController.java
@@ -11,13 +11,13 @@ import ru.practicum.compilation.CompilationClient;
 import ru.practicum.dto.CompilationDto;
 import ru.practicum.dto.CompilationRespDto;
 import ru.practicum.ViewsLoader;
+import ru.practicum.model.ConstraintViolationException;
 import ru.practicum.model.Marker;
 
 import javax.validation.constraints.Positive;
 
 @Controller
 @AllArgsConstructor
-@Validated
 @RequestMapping(path = "/admin/compilations")
 public class AdminCompilationClientController {
     private final CompilationClient client;
@@ -36,13 +36,19 @@ public class AdminCompilationClientController {
     }
 
     @DeleteMapping("/{compId}")
-    public ResponseEntity<Object> delete(@PathVariable @Positive Long compId) {
+    public ResponseEntity<Object> delete(@PathVariable Long compId) {
+        if (compId < 1) {
+            throw new ConstraintViolationException("Id should be positive");
+        }
         return client.delete(compId);
     }
 
     @PatchMapping("/{compId}")
     public ResponseEntity<Object> patch(@PathVariable @Positive Long compId,
                              @RequestBody @Validated(Marker.Update.class) CompilationDto compilationDto) {
+        if (compId < 1) {
+            throw new ConstraintViolationException("Id should be positive");
+        }
         ResponseEntity<Object> resp = client.patch(compId, compilationDto);
         if (resp.getStatusCode() == HttpStatus.OK) {
             CompilationRespDto dto = objectMapper.convertValue(resp.getBody(), CompilationRespDto.class);
